requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohttp=3.8.1=py39h89e85a6_0
aiosignal=1.2.0=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=21.4.0=pyhd8ed1ab_0
blas=1.0=mkl
blinker=1.4=py_1
bottleneck=1.3.2=py39he3068b8_1
brotlipy=0.7.0=py39h9ed2024_1003
ca-certificates=2021.10.8=h033912b_0
cachetools=4.2.4=pyhd8ed1ab_0
certifi=2021.10.8=py39h6e9494a_1
cffi=1.15.0=py39hc55c11b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.3=py39h6e9494a_1
cryptography=36.0.0=py39hf6deb26_0
frozenlist=1.3.0=py39h89e85a6_0
google-auth=2.6.0=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
gspread=4.0.1=pyh6c4a22f_0
idna=3.3=pyhd3eb1b0_0
intel-openmp=2021.4.0=hecd8cb5_3538
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.1=py39h4ab4a9b_0
mkl_random=1.2.2=py39hb2f4e1b_0
multidict=6.0.2=py39h89e85a6_0
ncurses=6.3=hca72f7f_2
numexpr=2.8.1=py39h2e5f0a9_0
numpy=1.21.2=py39h4b4dc7a_0
numpy-base=1.21.2=py39he0bd621_0
oauthlib=3.2.0=pyhd8ed1ab_0
openssl=1.1.1l=h0d85af4_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py39h743cdd8_0
pip=21.2.4=py39hecd8cb5_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd3eb1b0_0
pyjwt=2.3.0=pyhd8ed1ab_1
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.7=h88f2d9e_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=0.19.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2021.3=pyhd3eb1b0_0
pyu2f=0.1.5=pyhd8ed1ab_0
readline=8.1.2=hca72f7f_1
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.8=pyhd8ed1ab_0
setuptools=58.0.4=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.37.0=h707629a_0
tk=8.6.11=h7bc2e8c_0
typing-extensions=4.0.1=hd8ed1ab_0
typing_extensions=4.0.1=pyha770c72_0
tzdata=2021e=hda174b7_0
urllib3=1.26.8=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
yarl=1.7.2=py39h89e85a6_1
zlib=1.2.11=h4dc903c_4
